Fix travis build.
When we require emacs 25.1 there, it doesn't make sense to test it
with emacs 24. So this one got removed.

Don't use git snapshot emacs because there are some problems with TLS
on CI systems. Don't care more for now.

There is more some cask problems with python so this perl one-liner is
added as workaround.